implement new time-context codon to carry time across sequence segments: zo

+++ src/main/org/excelsi/nausicaa/ca/Pattern.java
@@ -28,6 +28,6 @@ public interface Pattern {
         /** random */
         public Random r;
         /** time step */
-        public long time;
+        public int time;
     }
 }
